requirements.txt
aiogram==3.2.0
yt-dlp
python-dotenv